#83 cleaned up tests and hardhat tasks
