+++ packages/hardhat-ts/package.json
@@ -16,7 +16,8 @@
     "send": "hardhat send",
     "generate": "hardhat generate",
     "account": "hardhat account",
-    "verify": "hardhat etherscan-verify --api-key PSW8C433Q667DVEX5BCRMGNAH9FSGFZ7Q8",
+    "verify:etherscan": "hardhat etherscan-verify --api-key PSW8C433Q667DVEX5BCRMGNAH9FSGFZ7Q8",
+    "verify:tenderly": "hardhat tenderly-verify",
     "lint": "eslint --config ./.eslintrc --ignore-path ./.eslintignore ./deploy/**/*.ts ./tasks/**/*.ts ./test/**/*.ts ./helpers/**/*.ts",
     "solhint": "solhint ./contracts/**/*.sol",
     "format": "yarn prettier -w ./deploy/**/*.ts ./tasks/**/*.ts ./test/**/*.ts ./helpers/**/*.ts contracts/**/*.sol ",
@@ -44,12 +45,16 @@
     "@tenderly/hardhat-tenderly": "^1.0.13",
     "@typechain/ethers-v5": "^9.0.0",
     "@typechain/hardhat": "^4.0.0",
+    "@types/qrcode-terminal": "^0.12.0",
     "chalk": "^4.1.2",
     "ethereum-waffle": "^3.4.0",
     "ethers": "^5.5.4",
     "hardhat": "^2.8.4",
     "hardhat-deploy": "^0.9.29",
     "hardhat-deploy-ethers": "^0.3.0-beta.13",
+    "hardhat-preprocessor": "^0.1.4",
+    "hardhat-storage-layout": "^0.1.6",
+    "hardhat-watcher": "^2.1.1",
     "ipfs-http-client": "^54.0.2",
     "qrcode-terminal": "^0.12.0",
     "scaffold-common": "^1.0.0",
@@ -61,6 +66,7 @@
     "@types/node": "^16.11.25",
     "@typescript-eslint/eslint-plugin": "^5.12.0",
     "@typescript-eslint/parser": "^5.12.0",
+    "bip39": "^3.0.4",
     "chai": "^4.3.6",
     "dotenv": "^14.3.2",
     "eslint": "^8.9.0",
@@ -72,6 +78,8 @@
     "eslint-plugin-mocha": "^10.0.3",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prettier": "^4.0.0",
+    "ethereumjs-util": "^7.1.4",
+    "ethereumjs-wallet": "^1.0.2",
     "glob": "^7.2.0",
     "mocha": "^9.2.1",
     "node-watch": "^0.7.3",
