+++ packages/hardhat-ts/tasks/helpers/debug.ts
@@ -1,0 +1,7 @@
+export const DEBUG = false;
+
+export const debugLog = (...args: any[]): void => {
+  if (DEBUG) {
+    console.log(args);
+  }
+};
