+++ packages/hardhat-ts/tasks/network.ts
@@ -1,0 +1,6 @@
+import { task } from 'hardhat/config';
+
+task('blockNumber', 'Prints the block number', async (_, { ethers }) => {
+  const blockNumber = await ethers.provider.getBlockNumber();
+  console.log(blockNumber);
+});
