+++ packages/vite-app-ts/src/config/appContracts.config.ts
@@ -11,8 +11,8 @@ import * as externalContracts from '~~/generated/external-contracts/esm/types';
 /**
  * ⛳️⛳️⛳️⛳️⛳️⛳️⛳️⛳️⛳️⛳️⛳️⛳️⛳️⛳️
  * ### Instructions
- * 1. edit externalContractsConfig.ts to add your external contract addresses.
- * 2. edit `contractConnectorConfig` function below and add them to the list
+ * 1. edit externalContracts.config.ts to add your external contract addresses.
+ * 2. edit `appContractsConfig` function below and add them to the list
  * 3. run `yarn contracts:build` to generate types for contracts
  * 4. run `yarn deploy` to generate hardhat_contracts.json
  *
@@ -23,17 +23,23 @@ import * as externalContracts from '~~/generated/external-contracts/esm/types';
 export const appContractsConfig = () => {
   try {
     const result = {
+      // --------------------------------------------------
       // 🙋🏽‍♂️ Add your hadrdhat contracts here
+      // --------------------------------------------------
       YourContract: createConnectorForHardhatContract(
         'YourContract',
         hardhatContracts.YourContract__factory,
         hardhatContractsJson
       ),
 
+      // --------------------------------------------------
       // 🙋🏽‍♂️ Add your external contracts here, make sure to define the address in `externalContractsConfig.ts`
+      // --------------------------------------------------
       DAI: createConnectorForExternalContract('DAI', externalContracts.DAI__factory, externalContractsAddressMap),
 
+      // --------------------------------------------------
       // 🙋🏽‍♂️ Add your external abi here (unverified contracts)`
+      // --------------------------------------------------
       // YourContract: createConnectorForExternalAbi(
       //   'YourContract',
       //   {
@@ -50,7 +56,7 @@ export const appContractsConfig = () => {
     return result;
   } catch (e) {
     console.error(
-      '❌ contractConnectorConfig: ERROR with loading contracts please run `yarn contracts:build or yarn contracts:rebuild`.  Then run `yarn deploy`!',
+      '❌ appContractsConfig: ERROR with loading contracts please run `yarn contracts:build or yarn contracts:rebuild`.  Then run `yarn deploy`!',
       e
     );
   }
